Move SignalMixin to common_py3.py.
